[Mole] Add create and update functions for Mole entity
In this commit I made mole (of MoleEntity class) a member of the GameBoard class.
The GameBoard now can create a mole at the start and move it around on the screen
randomly in equal time intervals. The mole has collision settings carried over
from the Obstacle in the Flap Bird example.